Restructure thigns a bit
Merge the Conclusiones and  Resultados sections.

Remove the metodology section as such, and split it into several smaller
sections. Including a new section for the explanation of the PSNR
algorithm.

Remove any results analysis made in the metodology section, and relocate
it to the results section.

Add the description of sections to follow in the introduction.

To whomever is reading this: Great going following the guidelines.